requirements.txt
argcomplete==3.2.1
asgiref==3.7.2
attrs==23.1.0
awscli==1.32.21
black==23.9.1
boto3==1.34.21
botocore==1.34.21
certifi==2023.7.22
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.6
django-allauth==0.52.0
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docutils==0.16
drf-spectacular==0.26.5
drf-standardized-errors==0.12.6
durationpy==0.6
hjson==3.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kappa==0.6.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
pipreqs==0.4.13
placebo==0.9.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
rsa==4.7.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
troposphere==4.5.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
whitenoise==6.6.0
yarg==0.1.9
cloudinary